- Refactor - Applied Strategy Pattern more DI and Things to better decoupling to all the chain of calls between viewModels and the netwrokManager

+++ RecipeManager/RecipeManager/Model/Recipe.swift
@@ -12,7 +12,7 @@ struct RecipesApiResponse: Decodable {
 }
 
 struct Recipe: Decodable {
-    let id: Float
+    let id: Int
     let title: String
     let image: String
